Refactor Pools
