+++ ui/core/src/api/MarketService/index.ts
@@ -1,5 +1,5 @@
 import { RWN } from "../../constants";
-import { Asset, AssetAmount, Pair } from "../../entities";
+import { Asset, AssetAmount, Pool } from "../../entities";
 import { SifUnSignedClient } from "../SifService/SifClient";
 
 export type MarketServiceContext = {
@@ -33,7 +33,7 @@ export default function createMarketService({
   sifApiUrl,
 }: MarketServiceContext) {
   const sifClient = new SifUnSignedClient(sifApiUrl);
-  const pairs = new Map<string, Pair>();
+  const poolMap = new Map<string, Pool>();
 
   async function generatePairs() {
     await loadAssets();
@@ -41,7 +41,7 @@ export default function createMarketService({
     return pools.map((poolData) => {
       const externalAssetTicker = poolData.external_asset.ticker;
 
-      const pair = Pair(
+      const pair = Pool(
         AssetAmount(RWN, poolData.native_asset_balance),
         AssetAmount(
           Asset.get(externalAssetTicker),
@@ -49,7 +49,7 @@ export default function createMarketService({
         )
       );
 
-      pairs.set(pair.symbol(), pair);
+      poolMap.set(pair.symbol(), pair);
     });
   }
 
@@ -59,7 +59,7 @@ export default function createMarketService({
     find(asset1: Asset | string, asset2: Asset | string) {
       if (!pairsGenerated.isResolved()) return null;
       const key = [asset1, asset2].map(toAssetSymbol).join("_");
-      return pairs.get(key) ?? null;
+      return poolMap.get(key) ?? null;
     },
   };
 }
