+++ ui/core/src/hooks/utils.test.ts
@@ -1,4 +1,4 @@
-import { AssetAmount, Pair, Coin, Network } from "../entities";
+import { AssetAmount, Pool, Coin, Network } from "../entities";
 import { assetPriceMessage } from "./utils";
 
 describe("assets with decimals", () => {
@@ -19,7 +19,7 @@ describe("assets with decimals", () => {
   test("assetPriceMessage", () => {
     const msg = assetPriceMessage(
       AssetAmount(ASSETS.atk, "100"),
-      Pair(
+      Pool(
         AssetAmount(ASSETS.atk, "1000000"),
         AssetAmount(ASSETS.btk, "2000000")
       ),
@@ -32,7 +32,7 @@ describe("assets with decimals", () => {
   test("with zero amounts message should be nothing", () => {
     const msg = assetPriceMessage(
       AssetAmount(ASSETS.atk, "0"),
-      Pair(AssetAmount(ASSETS.atk, "1"), AssetAmount(ASSETS.btk, "1")),
+      Pool(AssetAmount(ASSETS.atk, "1"), AssetAmount(ASSETS.btk, "1")),
       4
     );
     expect(msg).toBe("");
@@ -56,7 +56,7 @@ describe("assets with zero decimals", () => {
   test("with 1 as an amount", () => {
     const msg = assetPriceMessage(
       AssetAmount(ASSETS.atk, "1"),
-      Pair(
+      Pool(
         AssetAmount(ASSETS.atk, "1000000"),
         AssetAmount(ASSETS.btk, "1000000")
       )
@@ -67,7 +67,7 @@ describe("assets with zero decimals", () => {
   test("with 12 as an amount", () => {
     const msg = assetPriceMessage(
       AssetAmount(ASSETS.atk, "12"),
-      Pair(
+      Pool(
         AssetAmount(ASSETS.atk, "1000000"),
         AssetAmount(ASSETS.btk, "1000000")
       ),
